move CoinList from rewards helper to generic helpers
