+++ smart-contracts/contracts/cl-vault/src/vault/autocompound.rs
@@ -12,12 +12,13 @@ use osmosis_std::types::osmosis::concentratedliquidity::v1beta1::MsgCreatePositi
 use crate::helpers::{get_unused_balances, must_pay_one_or_two_from_balance};
 use crate::msg::{ExecuteMsg, MergePositionMsg};
 use crate::reply::Replies;
-use crate::rewards::CoinList;
 #[allow(deprecated)]
 use crate::state::USER_REWARDS;
 use crate::state::{MigrationStatus, MIGRATION_STATUS, POOL_CONFIG, POSITION};
 use crate::vault::concentrated_liquidity::create_position;
 use crate::ContractError;
+
+use super::helpers::CoinList;
 
 #[cw_serde]
 pub struct SwapAsset {
