+++ smart-contracts/contracts/cl-vault/src/vault/swap.rs
@@ -1,24 +1,11 @@
-use apollo_cw_asset::AssetInfo;
-use cosmwasm_std::{
-    to_json_binary, Addr, Coin, CosmosMsg, DepsMut, Env, Fraction, MessageInfo, Response, Uint128,
-    WasmMsg,
-};
-use cw_dex_router::{
-    msg::{
-        BestPathForPairResponse, ExecuteMsg as DexRouterExecuteMsg, QueryMsg as DexRouterQueryMsg,
-    },
-    operations::SwapOperationsListUnchecked,
-};
-use osmosis_std::types::{
-    cosmos::base::v1beta1::Coin as OsmoCoin, osmosis::poolmanager::v1beta1::SwapAmountInRoute,
-};
+use cosmwasm_std::{CosmosMsg, DepsMut, Env, Fraction, MessageInfo, Response, Uint128};
+use cw_dex_router::operations::SwapOperationsListUnchecked;
 
-use crate::state::DEX_ROUTER;
 use crate::state::POOL_CONFIG;
 use crate::{state::VAULT_CONFIG, ContractError};
 
-use super::autocompound::SwapAsset;
 use super::range::assert_range_admin;
+use super::{autocompound::SwapAsset, helpers::swap_msg};
 
 /// SwapCalculationResult holds the result of a swap calculation
 pub struct SwapCalculationResult {
@@ -131,216 +118,6 @@ pub fn execute_swap_non_vault_funds(
         .add_messages(swap_msgs)
         .add_attribute("method", "execute")
         .add_attribute("action", "swap_non_vault_funds"))
-}
-
-/// swap will always swap over the CL pool. In the future we may expand the
-/// feature such that it chooses best swaps over all routes
-pub fn swap_msg(deps: &DepsMut, env: &Env, params: SwapParams) -> Result<CosmosMsg, ContractError> {
-    // let pool_config = POOL_CONFIG.load(deps.storage)?;
-    let dex_router = DEX_ROUTER.may_load(deps.storage)?;
-
-    // we will only ever have a route length of one, this will likely change once we start selecting different routes
-    let pool_route = SwapAmountInRoute {
-        pool_id: params.pool_id,
-        token_out_denom: params.token_out_denom.to_string(),
-    };
-
-    // if we don't have a dex_router, we will always swap over the osmosis pool
-    if dex_router.is_none() {
-        return Ok(osmosis_swap_exact_amount_in_msg(
-            env,
-            pool_route,
-            params.token_in_amount,
-            &params.token_in_denom.to_string(),
-            params.token_out_min_amount,
-        ));
-    }
-
-    // we know we have a dex_router, so we can unwrap
-    let dex_router_address = dex_router.clone().unwrap();
-    let offer_asset = AssetInfo::Native(params.token_in_denom.to_string());
-    let ask_asset = AssetInfo::Native(params.token_out_denom.to_string());
-
-    let recommended_out: Uint128 = match params.recommended_swap_route.clone() {
-        Some(operations) => deps.querier.query_wasm_smart(
-            dex_router_address.to_string(),
-            &DexRouterQueryMsg::SimulateSwapOperations {
-                offer_amount: params.token_in_amount,
-                operations,
-            },
-        )?,
-        None => 0u128.into(),
-    };
-    let best_path: Option<BestPathForPairResponse> = deps.querier.query_wasm_smart(
-        dex_router_address.to_string(),
-        &DexRouterQueryMsg::BestPathForPair {
-            offer_asset: offer_asset.into(),
-            ask_asset: ask_asset.into(),
-            exclude_paths: None,
-            offer_amount: params.token_in_amount,
-        },
-    )?;
-    let best_out = match best_path.clone() {
-        Some(best_path) => best_path.return_amount,
-        None => 0u128.into(),
-    };
-
-    // if we need to force the route
-    if params.force_swap_route {
-        match params.recommended_swap_route {
-            Some(recommended_swap_route) => cw_dex_execute_swap_operations_msg(
-                &dex_router_address,
-                recommended_swap_route,
-                params.token_out_min_amount,
-                &params.token_in_denom.to_string(),
-                params.token_in_amount,
-            ),
-            None => Err(ContractError::TryForceRouteWithoutRecommendedSwapRoute {}),
-        }
-    } else if best_out.is_zero() && recommended_out.is_zero() {
-        Ok(osmosis_swap_exact_amount_in_msg(
-            env,
-            pool_route,
-            params.token_in_amount,
-            &params.token_in_denom.to_string(),
-            params.token_out_min_amount,
-        ))
-    } else if best_out.ge(&recommended_out) {
-        let operations = best_path
-            .ok_or(ContractError::MissingBestPath {})?
-            .operations
-            .into();
-        cw_dex_execute_swap_operations_msg(
-            &dex_router_address,
-            operations,
-            params.token_out_min_amount,
-            &params.token_in_denom.to_string(),
-            params.token_in_amount,
-        )
-    } else {
-        // recommended_out > best_out
-        let recommended_swap_route = params
-            .recommended_swap_route
-            .ok_or(ContractError::MissingRecommendedSwapRoute {})?;
-        cw_dex_execute_swap_operations_msg(
-            &dex_router_address,
-            recommended_swap_route, // will be some here
-            params.token_out_min_amount,
-            &params.token_in_denom.to_string(),
-            params.token_in_amount,
-        )
-    }
-}
-
-// fn find_optimal_swap_operation_route(
-//     querier: QuerierWrapper,
-//     dex_router_addr: &Addr,
-//     current_swap_route: &Option<SwapOperationsListUnchecked>, // TODO: Rename to recommended_swap_route
-//     token_in_denom: &String,
-//     token_in_amount: Uint128,
-//     token_out_denom: &String,
-//     force_swap_route: bool, // this true wants Some(current_swap_route)
-// ) -> Result<CosmosMsg, ContractError> {
-//     let offer_asset = AssetInfo::Native(token_in_denom.to_string());
-//     let ask_asset = AssetInfo::Native(token_out_denom.to_string());
-
-//     // recommended out it computed based on the simulation of the swap operations we perform against the cw-dex-router
-//     // using the recommended_swap_route
-//     let recommended_out: Uint128 = match current_swap_route.clone() {
-//         Some(operations) => querier.query_wasm_smart(
-//             dex_router_addr.to_string(),
-//             &DexRouterQueryMsg::SimulateSwapOperations {
-//                 offer_amount: token_in_amount,
-//                 operations,
-//             },
-//         )?,
-//         None => Uint128::zero(),
-//     };
-
-//     // best_path is retrieven by the iteration on the cw-dex-router using all the paths available on its state
-//     // and picking up the one with higher token_out_amount
-//     let best_path: Option<BestPathForPairResponse> = querier.query_wasm_smart(
-//         dex_router_addr.to_string(),
-//         &DexRouterQueryMsg::BestPathForPair {
-//             offer_asset: offer_asset.into(),
-//             ask_asset: ask_asset.into(),
-//             exclude_paths: None,
-//             offer_amount: token_in_amount,
-//         },
-//     )?;
-//     let best_outcome = best_path
-//         .as_ref()
-//         .map_or(Uint128::zero(), |path| path.return_amount);
-
-//     // Determine the route to use
-//     let route = if force_swap_route {
-//         current_swap_route
-//             .clone()
-//             .ok_or(ContractError::TryForceRouteWithoutRecommendedSwapRoute {})?
-//     } else if best_outcome >= recommended_out {
-//         best_path.expect("Expected a best path").operations.into()
-//     } else {
-//         current_swap_route
-//             .clone()
-//             .expect("Expected a recommended route")
-//     };
-
-//     let token_out_min_amount = todo!();
-
-//     // TODO: If we reach that point with a 0 token_out_min_amount
-//     // which is lower than allowed by config.max_slippage throw an Error()
-
-//     cw_dex_execute_swap_operations_msg(
-//         dex_router_addr,
-//         route,
-//         token_out_min_amount,
-//         token_in_denom,
-//         token_in_amount,
-//     )
-// }
-
-fn osmosis_swap_exact_amount_in_msg(
-    env: &Env,
-    pool_route: SwapAmountInRoute,
-    token_in_amount: Uint128,
-    token_in_denom: &String,
-    token_out_min_amount: Uint128,
-) -> CosmosMsg {
-    osmosis_std::types::osmosis::poolmanager::v1beta1::MsgSwapExactAmountIn {
-        sender: env.contract.address.to_string(),
-        routes: vec![pool_route],
-        token_in: Some(OsmoCoin {
-            denom: token_in_denom.to_string(),
-            amount: token_in_amount.to_string(),
-        }),
-        token_out_min_amount: token_out_min_amount.to_string(),
-    }
-    .into()
-}
-
-fn cw_dex_execute_swap_operations_msg(
-    dex_router_address: &Addr,
-    operations: SwapOperationsListUnchecked,
-    token_out_min_amount: Uint128,
-    token_in_denom: &String,
-    token_in_amount: Uint128,
-) -> Result<CosmosMsg, ContractError> {
-    let swap_msg: CosmosMsg = WasmMsg::Execute {
-        contract_addr: dex_router_address.to_string(),
-        msg: to_json_binary(&DexRouterExecuteMsg::ExecuteSwapOperations {
-            operations,
-            minimum_receive: Some(token_out_min_amount),
-            to: None,
-            offer_amount: None,
-        })?,
-        funds: vec![Coin {
-            denom: token_in_denom.to_string(),
-            amount: token_in_amount,
-        }],
-    }
-    .into();
-
-    Ok(swap_msg)
 }
 
 /// estimate_swap can be used to pass correct token_out_min_amount values into swap()
